requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.2
Faker==4.14.0
gunicorn==20.0.4
mysql-connector-python==8.0.22
protobuf==3.14.0
pip==20.2.3
psycopg2==2.8.5
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
setuptools==50.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
wheel==0.35.1
whitenoise==5.2.0
wincertstore==0.2
